+++ packages/helper/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+  },
+  "include": ["**/src/**/*.ts"],
+  "exclude": ["**/src/**/*.test.ts"]
+}
