--- packages/schema-to-library/src/helper/index.ts
@@ -1,3 +1,0 @@
-export { resolveSchemaDependenciesFromSchema } from './dep/resolve-schema-dependencies-from-shema.js'
-export { type } from './dep/type.js'
-export { normalizeTypes, toPascalCase } from './utils/index.js'
