--- packages/schema-to-library/vitest.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'vitest/config'
-export default defineConfig({
-  test: {
-    includeSource: ['src/**/*.ts'],
-    include: ['src/**/*.test.ts'],
-    coverage: {
-      include: ['src/**/*.ts'],
-      all: true,
-    },
-  },
-})
