+++ packages/zod/src/index.ts
@@ -1,6 +1,6 @@
-import type { Schema } from '../../../cli/src/index.js'
-import { resolveSchemaDependenciesFromSchema, toPascalCase, type } from '../helper/index.js'
-import zod from './zod/index.js'
+import type { Schema } from '@schema-to-library/cli'
+import { resolveSchemaDependenciesFromSchema, toPascalCase, type } from '../../schema-to-library/src/helper/index.js'
+import zod from '../../schema-to-library/src/zod/zod/index.js'
 
 export function schemaToZod(schema: Schema): string {
   // Get the root schema name from title or use default
